requirements.txt
discord.py==0.16.12
numpy==1.15.4